requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
python-gitlab==3.12.0
requests==2.28.1
requests-toolbelt==0.10.1
sh==1.14.3
smmap==5.0.0
urllib3==1.26.13
Werkzeug==2.2.2
